Extract code into separate class
